% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{guess_filename}
\alias{guess_filename}
\title{guess_filename}
\usage{
guess_filename(x)
}
\description{
guess_filename
}
\examples{
guess_filename("https://agupubs.onlinelibrary.wiley.com/doi/pdfdirect/10.1002/2016WR020175?download=true")

}
