% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscihub-package.R
\docType{package}
\name{rscihub}
\alias{rscihub}
\alias{_PACKAGE}
\alias{rscihub-package}
\title{rscihub}
\description{
Download papers batchly using DOIs from many database. And some practical functions.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{DOI}
\keyword{download}
\keyword{internal}
\keyword{paper}
