% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_FUN.R
\name{src_URL}
\alias{src_URL}
\alias{srcFUN}
\alias{url_AMS}
\alias{url_IOP}
\alias{url_hess}
\alias{url_springer}
\alias{url_nature}
\title{srcFUN}
\usage{
src_URL(url)

url_AMS(doi)

url_IOP(doi)

url_hess(doi)

url_springer(doi)

url_nature(doi)
}
\arguments{
\item{url}{pdf url}

\item{doi}{Character, Digital Object Identifier, like
"10.1175/JHM-D-15-0157.1", URLencoding format is also supported, i.e.
"10.1175\\%2FJHM-D-15-0157.1".
Based on doi, \code{srcFUN} find corresponding paper and download it.}

\item{urls}{pdf urls}

\item{outdir}{Output directory}

\item{.download}{If true, it will will download pdf directly, and return
pdf src. If false, only pdf src returned, without downlaoding pdf.}

\item{...}{other parameters pass to \code{\link[httr:GET]{httr::GET()}}}
}
\description{
\itemize{
\item \code{src_URL} simplest srcFUN, just treat the input url as download link.
\item \code{src_wiley} wiley library. Journal like GRL, JGR, WRR, HP,
all in the database.
}

If just src returned, you need to download with \code{\link[=download_httr]{download_httr()}}
}
